Update ideal_weight.rb
Тип переменно heights изменен на Float

+++ exercise_1/ideal_weight.rb
@@ -1,5 +1,5 @@
 print "Представьтесь, пожалуйста: "
 name = gets.chomp
 print "Какой у вас рост в сантиметрах? "
-heights = gets.chomp.to_i
+heights = gets.chomp.to_f
 puts "#{name}, ваш идеальный вес: #{heights-110} кг"
